+++ packages/components/CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.9.3-rc.1] - 2022-04-06
+
+#### Added
+- `Text Field Wrapper` now has a `showCharacterCount` property which can be used to hide the character count when a `maxLength` attribute is set on the wrapped `input`.
+- `Textarea Wrapper` now has a `showCharacterCount` property which can be used to hide the character count when a `maxLength` attribute is set on the wrapped `textarea`.
+
 ### [2.9.3-rc.0-skeletons] - 2022-03-29
 
 ### [2.9.3-rc.0] - 2022-03-28